+++ vimeo-networking/src/main/java/com/vimeo/networking/callers/GetRequestCaller.java
@@ -35,7 +35,7 @@ import com.vimeo.networking.model.ChannelList;
 import com.vimeo.networking.model.Comment;
 import com.vimeo.networking.model.CommentList;
 import com.vimeo.networking.model.FeedList;
-import com.vimeo.networking.model.LiveStats;
+import com.vimeo.networking.model.live.LiveStats;
 import com.vimeo.networking.model.RecommendationList;
 import com.vimeo.networking.model.User;
 import com.vimeo.networking.model.UserList;
