+++ vimeo-networking/src/test/java/com/vimeo/networking/model/live/LiveTimeTest.kt
@@ -22,20 +22,20 @@
  * SOFTWARE.
  */
 
-package com.vimeo.networking.model
+package com.vimeo.networking.model.live
 
 import com.vimeo.networking.Utils
 import org.junit.Test
 
 /**
- * Tests for the [Live] class
+ * Tests for the [LiveTime] class
  *
- * Created by zetterstromk on 8/23/17.
+ * Created by zetterstromk on 9/11/17.
  */
-class LiveTest {
+class LiveTimeTest {
 
     @Test
     fun verifyTypeAdapterWasGenerated() {
-        Utils.verifyTypeAdapterGeneration(Live::class.java)
+        Utils.verifyTypeAdapterGeneration(LiveTime::class.java)
     }
 }
